Add delete question/answer confirmation.

+++ src/components/shared/EditDelete.tsx
@@ -11,6 +11,17 @@ import { deleteQuestionById } from "@/server/actions/question.action";
 import { deleteAnswerById } from "@/server/actions/answer.action";
 import { useToast } from "@/components/ui/use-toast";
 import { capitalize } from "@/lib/format";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 interface EditDeleteProps {
   type: "question" | "answer";
@@ -42,7 +53,7 @@ export default function EditDelete({ type, typeId, clerkId }: EditDeleteProps) {
   return (
     <SignedIn>
       {pathname.startsWith("/profile/") && userId === clerkId && (
-        <div className="flex-start gap-3 pl-3">
+        <div className="flex-end min-w-[52px] gap-3 pl-3">
           {type === "question" && (
             <Link href={`/question/edit/${typeId}`}>
               <Button className="p-0">
@@ -55,26 +66,45 @@ export default function EditDelete({ type, typeId, clerkId }: EditDeleteProps) {
               </Button>
             </Link>
           )}
-          <Button
-            disabled={isLoading}
-            className="p-0"
-            onClick={() => handleDeleteAction(type)}
-          >
-            {isLoading ? (
-              <ReloadIcon
-                className="animate-spin text-red-500"
-                width={14}
-                height={14}
-              />
-            ) : (
-              <Image
-                src="/assets/icons/trash.svg"
-                alt="delete"
-                width={14}
-                height={14}
-              />
-            )}
-          </Button>
+          <AlertDialog>
+            <AlertDialogTrigger>
+              <Button disabled={isLoading} className="p-0">
+                {isLoading ? (
+                  <ReloadIcon
+                    className="animate-spin text-red-500"
+                    width={14}
+                    height={14}
+                  />
+                ) : (
+                  <Image
+                    src="/assets/icons/trash.svg"
+                    alt="delete"
+                    width={14}
+                    height={14}
+                  />
+                )}
+              </Button>
+            </AlertDialogTrigger>
+            <AlertDialogContent className="text-dark100_light900 background-light900_dark200">
+              <AlertDialogHeader>
+                <AlertDialogTitle>{`Delete ${type}?`}</AlertDialogTitle>
+                <AlertDialogDescription>
+                  {`This action cannot be undone. This will permanently delete your ${type}.`}
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel className="background-light800_dark400">
+                  Cancel
+                </AlertDialogCancel>
+                <AlertDialogAction
+                  className="bg-red-500 text-white"
+                  onClick={() => handleDeleteAction(type)}
+                >
+                  Delete
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       )}
     </SignedIn>
